+++ src/Projecte/Recurs.java
@@ -3,6 +3,6 @@ package Projecte;
 public interface Recurs {
 	String getNom();
 	String getTipus();
-	int puntsASumar();
+	double puntsASumar();
 
 }
